+++ src/WebJobs.Extensions.DurableTask/Grpc/LegacyLocalGrpcListener.cs
@@ -1,0 +1,123 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+#nullable enable
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Threading;
+using System.Threading.Tasks;
+using Grpc.Core;
+using P = Microsoft.DurableTask.Protobuf;
+
+namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Grpc
+{
+    internal class LegacyLocalGrpcListener : ILocalGrpcListener
+    {
+        private const int DefaultPort = 4001;
+
+        // Pick a large, fixed range of ports that are going to be valid in all environment.
+        // Avoiding ports below 1024 as those are blocked by app service sandbox.
+        // Ephemeral ports for most OS start well above 32768. See https://www.ncftp.com/ncftpd/doc/misc/ephemeral_ports.html
+        private const int MinPort = 30000;
+        private const int MaxPort = 31000;
+
+        private readonly DurableTaskExtension extension;
+
+        private readonly Random portGenerator;
+        private readonly HashSet<int> attemptedPorts;
+
+        private Server? grpcServer;
+
+        public LegacyLocalGrpcListener(DurableTaskExtension extension)
+        {
+            this.extension = extension ?? throw new ArgumentNullException(nameof(extension));
+
+            this.portGenerator = new Random();
+            this.attemptedPorts = new HashSet<int>();
+        }
+
+        public string? ListenAddress { get; private set; }
+
+        public async Task StartAsync(CancellationToken cancelToken)
+        {
+            const int maxAttempts = 10;
+            int numAttempts = 1;
+            while (numAttempts <= maxAttempts)
+            {
+                ChannelOption[] options = new[]
+                {
+                    new ChannelOption(ChannelOptions.MaxReceiveMessageLength, int.MaxValue),
+                    new ChannelOption(ChannelOptions.MaxSendMessageLength, int.MaxValue),
+                };
+
+                if (this.grpcServer != null)
+                {
+                    await this.grpcServer.ShutdownAsync();
+                }
+
+                this.grpcServer = new Server(options);
+                this.grpcServer.Services.Add(P.TaskHubSidecarService.BindService(new TaskHubGrpcServer(this.extension)));
+
+                int listeningPort = numAttempts == 1 ? DefaultPort : this.GetRandomPort();
+                int portBindingResult = this.grpcServer.Ports.Add("localhost", listeningPort, ServerCredentials.Insecure);
+                if (portBindingResult != 0)
+                {
+                    try
+                    {
+                        this.grpcServer.Start();
+                        this.ListenAddress = $"http://localhost:{listeningPort}";
+
+                        this.extension.TraceHelper.ExtensionInformationalEvent(
+                            this.extension.Options.HubName,
+                            instanceId: string.Empty,
+                            functionName: string.Empty,
+                            message: $"Opened local gRPC endpoint: {this.ListenAddress}",
+                            writeToUserLogs: true);
+
+                        return;
+                    }
+                    catch (IOException)
+                    {
+                        portBindingResult = 0;
+                    }
+                }
+
+                if (portBindingResult == 0)
+                {
+                    this.extension.TraceHelper.ExtensionWarningEvent(
+                        this.extension.Options.HubName,
+                        functionName: string.Empty,
+                        instanceId: string.Empty,
+                        message: $"Failed to open local port {listeningPort}. This was attempt #{numAttempts} to open a local port.");
+                    this.attemptedPorts.Add(listeningPort);
+                    numAttempts++;
+                }
+            }
+
+            throw new IOException($"Unable to find a port to open an RPC endpoint on after {maxAttempts} attempts");
+        }
+
+        public async Task StopAsync(CancellationToken cancelToken)
+        {
+            if (this.grpcServer != null)
+            {
+                await this.grpcServer.ShutdownAsync();
+            }
+        }
+
+        private int GetRandomPort()
+        {
+            // Get a random port that has not already been attempted so we don't waste time trying
+            // to listen to a port we know is not free.
+            int randomPort;
+            do
+            {
+                randomPort = this.portGenerator.Next(MinPort, MaxPort);
+            }
+            while (this.attemptedPorts.Contains(randomPort));
+
+            return randomPort;
+        }
+    }
+}
