+++ src/WebJobs.Extensions.DurableTask/Grpc/Protos/refresh-protos.ps1
