+++ src/WebJobs.Extensions.DurableTask/WebJobs.Extensions.DurableTask.csproj
@@ -28,7 +28,6 @@
     <Version>$(MajorVersion).$(MinorVersion).$(PatchVersion)-$(VersionSuffix)</Version>
   </PropertyGroup>
 
-
   <!-- NuGet Publishing Metadata -->
   <PropertyGroup>
     <Title>Azure Functions Durable Task Extension</Title>
@@ -50,7 +49,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Protobuf Include="**/*.proto" GrpcServices="Both" Access="Internal"  />
+    <Protobuf Include="Grpc/Protos/**/*.proto" GrpcServices="Both" Access="Internal"  />
   </ItemGroup>
 
   <ItemGroup>
